backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
build==1.0.3
CacheControl==0.13.1
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.8
Django==4.2.10
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
installer==0.7.0
isort==5.13.2
jaraco.classes==3.3.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.17.1
shellingham==1.5.4
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2024.1.31
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
virtualenv==20.25.0
zipp==3.17.0
